1. Extract RunLoad from main.go 2. Add API Server at server.go 3. Add route to slowcooker entry
