+++ route/response.go
@@ -1,0 +1,6 @@
+package route
+
+// Response : response message
+type Response struct {
+	status, message string
+}
